requirements.txt
# 私人金融分析师后端项目依赖

# ===== 核心Web框架 =====
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# ===== 数据库相关 =====
sqlalchemy>=2.0.0
pymysql>=1.1.0
alembic>=1.12.0

# ===== 认证和安全 =====
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# ===== 数据验证 =====
pydantic>=2.0.0
pydantic-settings>=2.0.0

# ===== LangChain AI框架 =====
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.0.20
langchain-experimental>=0.0.40

# ===== Tavily搜索 =====
langchain-tavily>=0.1.0

# ===== OpenAI =====
openai>=1.0.0

# ===== 数据处理 =====
pandas>=2.0.0
numpy>=1.24.0

# ===== HTTP客户端 =====
requests>=2.31.0
aiohttp>=3.8.0
httpx>=0.25.0

# ===== 时间处理 =====
python-dateutil>=2.8.0

# ===== 配置文件处理 =====
pyyaml>=6.0

# ===== 日志 =====
loguru>=0.7.0

# ===== 进度条 =====
tqdm>=4.65.0

# ===== 异步支持 =====
aiofiles>=23.0.0

# ===== 数据可视化（可选） =====
matplotlib>=3.7.0
seaborn>=0.12.0

# ===== 开发工具（可选） =====
jupyter>=1.0.0
ipython>=8.0.0

# ===== 测试工具（可选） =====
pytest>=7.0.0
pytest-asyncio>=0.21.0